updated event to plan + central GET planParameters
updated event to plan + central GET planParameters
